// frontend/src/app/core/api/model/asset-add-request.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AssetAddRequest { 
    assetType: AssetAddRequest.AssetTypeEnum;
    symbol: string;
    name: string;
    currencyCode: string;
    exchangeId?: string;
    companyId?: string;
    fundManager?: string;
    assetClass?: AssetAddRequest.AssetClassEnum;
    trackedIndexId?: string;
    aliases?: { [key: string]: string; };
}
export namespace AssetAddRequest {
    export const AssetTypeEnum = {
        Stock: 'STOCK',
        Etf: 'ETF',
        Index: 'INDEX'
    } as const;
    export type AssetTypeEnum = typeof AssetTypeEnum[keyof typeof AssetTypeEnum];
    export const AssetClassEnum = {
        Equity: 'EQUITY',
        Bond: 'BOND',
        Commodity: 'COMMODITY',
        Mixed: 'MIXED'
    } as const;
    export type AssetClassEnum = typeof AssetClassEnum[keyof typeof AssetClassEnum];
}