// frontend/src/app/core/api/model/currency-response.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CurrencyResponse { 
    code?: string;
    symbol?: string;
    name?: string;
}